Output each entry as a link. Add more robust output handling.
Each link contains href=# at this point. My plan is to either:

Generate some kind of pointer with each node indicating the directory
that the command was executed from, thereby yeilding a way of computing
a filepath, or

generate the filepaths when the node is built and add an attribute to
each a tag indicating the absolute path to the file.
e Changes to be committed: