Output each command to a div with a unique id.
The runCmd function no longer accepts a callback for handling output,
it is all handled internally.

First the command output is initialized as an html div node containing
the command and any args passed to it.

This node is then passed as a jquery context each time the child_process
returns any data from the stdout pipe, which we wrap in html and append
to the context.