Move the command inoput field into AutocompleteComponent.
It was difficult to update the input field as a parent of the
Autocomplete component.  I imagined two options:

1) Use Flux pattern to store the autocompletes and update the field as
nessicary.

2) Broadcast a CustomEvent when the Autocomplete updates and listen to
it on the CommandLine component.  This is an anti-pattern in Flux and
the very thing it tries to avoid with uni-directional data flow.

Since option 1 seems like overkill with no immediate benefit
and option 2 being an antipattern I opted to just build the input into
the AutocompleteComponent.